backend/requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
python-dotenv>=1.0.0,<1.1.0
Pillow>=9.5.0,<9.6.0
requests>=2.31.0,<2.32.0
pyjwt==2.8.0
cryptography==41.0.5
psutil==7.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise>=6.5.0,<6.6.0
django-cors-headers>=4.3.0,<4.4.0
beautifulsoup4==4.12.2
bs4==0.0.1
google-genai>=1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
rembg>=2.0.50
onnxruntime>=1.15.0
